Moved SortingStuff class to Classes folder and restructured the class
